Entity User added. And also added next dependencies: 1. Added import Client from old System. 2. Added import Notes from old System. 3. Added tables connection. 4. Table patient full. 5. Added one to many connection. 6. Added dependency in application.properties and etc.
