+++ pom.xml
@@ -12,7 +12,7 @@
 
     <groupId>org.cleverdev.clientservice</groupId>
     <artifactId>api-for-cleverdev</artifactId>
-    <version>api-new-system-v0.3</version>
+    <version>api-new-system-v0.4</version>
     <name>cleverdevtest-client-service</name>
     <description>cleverdevtest-client-service</description>
