+++ src/main/java/com/cleverdev/clientService/CleverDevNewSystem.java
@@ -1,15 +1,10 @@
 package com.cleverdev.clientService;
 
-import com.cleverdev.clientService.patient.controller.PatientControllerNewSystem;
-import com.google.gson.Gson;
+import com.cleverdev.clientService.controller.NoteController;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.domain.Page;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.util.List;
 
 /**
  * Created by Vladislav Domaniewski
@@ -19,11 +14,7 @@ import java.util.List;
 public class CleverDevNewSystem {
 
     public static void main(String[] args) throws Exception {
-
         SpringApplication.run(CleverDevNewSystem.class, args);
-        PatientControllerNewSystem patien = new PatientControllerNewSystem();
-
-        patien.getJsonFromOldSystem();
     }
 
 }
