+++ src/main/java/com/cleverdev/clientService/controller/NoteController.java
@@ -1,0 +1,19 @@
+package com.cleverdev.clientService.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.java.Log;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+/**
+ * Created by Vladislav Domaniewski
+ */
+
+@Log
+@RestController
+@RequestMapping("/api")
+@RequiredArgsConstructor
+public class NoteController {
+
+}
