+++ src/main/java/com/cleverdev/clientService/converter/PatientConvert.java
@@ -1,0 +1,12 @@
+package com.cleverdev.clientService.converter;
+
+import org.springframework.stereotype.Component;
+
+/**
+ * Created by Vladislav Domaniewski
+ */
+
+@Component
+public class PatientConvert {
+
+}
