+++ src/main/java/com/cleverdev/clientService/core/DataException.java
@@ -1,0 +1,8 @@
+package com.cleverdev.clientService.core;
+
+/**
+ * Created by Vladislav Domaniewski
+ */
+
+public class DataException {
+}
