+++ src/main/java/com/cleverdev/clientService/enums/PatientStatusEnum.java
@@ -1,0 +1,11 @@
+package com.cleverdev.clientService.enums;
+
+/**
+ * Created by Vladislav Domaniewski
+ */
+
+public enum PatientStatusEnum {
+    PENDING,
+    ACTIVE,
+    INACTIVE
+}
