--- src/main/java/com/cleverdev/clientService/note/dto/PatientNote.java
@@ -1,8 +1,0 @@
-package com.cleverdev.clientService.note.dto;
-
-/**
- * Created by Vladislav Domaniewski
- */
-
-public class PatientNote {
-}
