--- src/main/java/com/cleverdev/clientService/patient/entity/Patient.java
@@ -1,42 +1,0 @@
-package com.cleverdev.clientService.patient.entity;
-
-import com.cleverdev.clientService.patient.enums.PatientStatusEnum;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-/**
- * Created by Vladislav Domaniewski
- */
-
-@Entity
-@Data
-@Table(name = "patient_profile")
-@NoArgsConstructor
-public class Patient {
-    public Patient(String firstName, String lastName, String oldClientGuid, PatientStatusEnum statusId) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.oldClientGuid = oldClientGuid;
-        this.statusId = statusId;
-    }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-    @Column(name = "old_client_guid")
-    private String oldClientGuid;
-
-    @Column(name = "status_id")
-    private PatientStatusEnum statusId;
-    // value from Old version
-//    private String agency;
-}
