--- src/main/java/com/cleverdev/clientService/patient/enums/PatientStatusEnum.java
@@ -1,7 +1,0 @@
-package com.cleverdev.clientService.patient.enums;
-
-public enum PatientStatusEnum {
-    PENDING,
-    ACTIVE,
-    INACTIVE,
-}
