--- src/main/java/com/cleverdev/clientService/patient/repository/PatientRepository.java
@@ -1,9 +1,0 @@
-package com.cleverdev.clientService.patient.repository;
-
-
-import com.cleverdev.clientService.patient.entity.Patient;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PatientRepository extends JpaRepository<Patient, Long> {
-
-}
