--- src/main/java/com/cleverdev/clientService/patient/service/PatientService.java
@@ -1,5 +1,0 @@
-package com.cleverdev.clientService.patient.service;
-
-public interface PatientService {
-    void populatingDatabase(String url) throws Exception;
-}
