--- src/main/java/com/cleverdev/clientService/patient/service/converter/PatientConvert.java
@@ -1,9 +1,0 @@
-package com.cleverdev.clientService.patient.service.converter;
-
-/**
- * Created by Vladislav Domaniewski
- */
-
-public class PatientConvert {
-
-}
