+++ src/main/java/com/cleverdev/clientService/repository/NoteRepository.java
@@ -1,0 +1,7 @@
+package com.cleverdev.clientService.repository;
+
+import com.cleverdev.clientService.entity.Note;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface NoteRepository extends JpaRepository<Note, Long> {
+}
