+++ src/main/java/com/cleverdev/clientService/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.cleverdev.clientService.repository;
+
+import com.cleverdev.clientService.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
