--- src/main/java/com/cleverdev/clientService/user/dto/UserDto.java
@@ -1,17 +1,0 @@
-package com.cleverdev.clientService.user.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * Created by Vladislav Domaniewski
- */
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserDto {
-    private Long id;
-    private String login;
-}
