+++ src/main/resources/application.properties
@@ -11,3 +11,18 @@ spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.proc.param_null_passing=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL94Dialect
+
+server.error.include-binding-errors=always
+server.error.include-message=always
+springdoc.api-docs.enabled=true
+springdoc.swagger-ui.enabled=true
+springdoc.swagger-ui.path=/swagger-ui
+springdoc.swagger-ui.disable-swagger-default-url=true
+springdoc.swagger-ui.display-operation-id=true
+springdoc.swagger-ui.display-request-duration=true
+springdoc.swagger-ui.show-common-extensions=true
+springdoc.swagger-ui.show-extensions=true
+
+#????? ????????? ?????? ??? ?????? ?? ?????? ????????
+app.urlClients = http://localhost:8081/clients
+app.urlNotes = http://localhost:8081/notes
