+++ src/test/java/notes/PatientControllerTest.java
@@ -1,7 +1,7 @@
 package notes;
 
-import com.cleverdev.clientService.patient.dto.ClientNoteRequestDto;
-import com.cleverdev.clientService.patient.repository.PatientRepository;
+import com.cleverdev.clientService.dto.ClientNoteRequestDto;
+import com.cleverdev.clientService.repository.PatientRepository;
 import lombok.RequiredArgsConstructor;
 import org.junit.After;
 import org.junit.Test;
@@ -9,8 +9,6 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import java.time.LocalDate;
